tests: deprecate test_tools; replace with tests.pytest_plugin
